D109: 200+ project improvements — attribute macros, MODULE metadata, kernel-doc, unused includes, param names, fallthrough, style fixes
Batch AB: __must_check, __malloc, __read_mostly, __cacheline_aligned, __no_sanitize_address macros in types.h
Batch C: MODULE_LICENSE/DESCRIPTION/AUTHOR in 5 module files + 23 kernel-doc comments (gpiolib.c, pm_runtime.c, tpm_attest.c, dynamic_debug.c)
Batch D: 9 unused heap.h includes removed + 9 function pointer param names added + TODO resolved + 6 style fixes + format warning suppression for kernel's custom size_t=uint64_t

~90 files changed
Zero-warning build with -Wall -Wextra -Werror

+++ src/drivers/nbd.c
@@ -268,7 +268,7 @@ void nbd_disconnect(int dev_id)
     blockdev_unregister(dev_id);
     net_tcp_close(dev->conn_id);
     kprintf("[nbd] Device nbd%d (id=%d) disconnected\n",
-            dev - g_nbd_devices, dev_id);
+            (int)(dev - g_nbd_devices), dev_id);
     memset(dev, 0, sizeof(*dev));
 }
